requirements.txt
# Core audio
sounddevice>=0.5.0
numpy>=1.24.0

# STT Engines
faster-whisper>=1.1.0
openai>=1.30.0

# GUI
customtkinter>=5.2.0
pillow>=10.0.0

# System Integration
pystray>=0.19.0
keyboard>=0.13.5
mouse>=0.7.1

# LLM for text normalization
google-genai>=1.0.0
